황정연 - 홍보 등록&리스트

+++ ecoshop/src/main/java/com/sp/app/admin/model/PromotionManage.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class PromotionManage {
 	private long promotionId;
 	private long memberId;
-	private long advertisingId;
+	private Integer advertisingId;
 	private String subject;
 	private String imageFilename;
 	private String regDate;
